requirements.txt
numpy
pandas
scikit-learn
prophet
keras
tensorflow
grpcio
grpcio-tools
protobuf
prometheus-api-client